+++ Tests/Apps/NewLocaleCompanion/NewLocaleCompanion/NewLocaleCompanionApp.swift
@@ -1,0 +1,18 @@
+//
+//  NewLocaleCompanionApp.swift
+//  NewLocaleCompanion
+//
+//  Created by Daniel Saidi on 2024-12-19.
+//
+
+import SwiftUI
+
+@main
+struct NewLocaleCompanionApp: App {
+    
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}